refactor: rename Hours component
