+++ components/Hours/index.ts
@@ -1,0 +1,2 @@
+export { default as Hours } from "./Hours";
+export * from "./utils";
