+++ components/Hours/utils.ts
